practica-2-f1-main/requirements.txt
alembic==1.7.7
aniso8601==9.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.0
colorama==0.4.4
Flask==2.1.0
Flask-Cors==3.0.10
Flask-HTTPAuth==4.6.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
passlib==1.7.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
urllib3==1.26.9
Werkzeug==2.1.0
zipp==3.7.0
